refactor: extract character management logic into dedicated classes
Extract character list management, character details management, and character creation logic from monolithic event handlers into dedicated classes (CharListManager, CharacterManager, CharCreationManager). This improves code organization, separation of concerns, and maintainability. The refactoring centralizes related functionality, reduces coupling between components, and makes the codebase more modular and testable.
